backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.27.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
PyMuPDF==1.24.3
faiss-cpu==1.7.4
numpy==1.26.4
httpx==0.27.0
google-cloud-aiplatform==1.49.0
slowapi==0.1.9
python-multipart==0.0.9
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
